redirecting to book after adding
