+++ src/pages/user/AddBookPage/AddBook.jsx
@@ -1,7 +1,7 @@
 import { Formik, Form } from "formik";
 import * as Yup from "yup";
-import { useCallback, useContext, useEffect } from "react";
-import { useLocation } from "react-router-dom";
+import { useCallback, useContext } from "react";
+import { useLocation, useNavigate } from "react-router-dom";
 import BookContext from "../../../context/BookContext";
 import TextField from "./components/TextField";
 import RadioButton from "./components/RadioButton";
@@ -10,7 +10,7 @@ import CheckboxField from "./components/CheckboxField";
 import TextareaField from "./components/TextareaField";
 import DateElement from "./components/DateElement";
 import booksService from '../../../services/books';
-import ScrollToError from "./components/ScrollToError";
+import ScrollToTop from "./components/ScrollToTop";
 
 const moodOptions = [
     { label: 'In love', value: 'in_love' },
@@ -26,7 +26,8 @@ const moodOptions = [
 const AddBook = () => {
     const { state } = useLocation()
     const { refreshBooks } = useContext(BookContext)
-
+    const navigate = useNavigate()
+    const location = useLocation()
 
     const initialValues = {
         title: state?.title || '',
@@ -47,8 +48,6 @@ const AddBook = () => {
     })
 
     const onSubmit = useCallback(async (values, { resetForm, setStatus, setValues }) => {
-        // setFieldTouched('title')
-        // setFieldTouched('author')
         const moodsPercentages = {}
         for (const [key, value] of Object.entries(values.moodsrate)) {
             if (values.status === "read" && values.moods.includes(key)) {
@@ -87,10 +86,12 @@ const AddBook = () => {
 
         const bookData = values.status === "read" ? readBook : values.status === "reading" ? readingBook : toReadBook
 
-        const el = document.querySelector('#top')
-
         try {
             await booksService.addBook(bookData)
+            const books = await booksService.getBooks()
+            const bookFiltered = books[bookData.status].filter(book => book.title === bookData.title && book.author === bookData.author)[0]
+            console.log('book filtered', bookFiltered)
+            navigate(`/books/${bookFiltered.id}`, { state: location.pathname })
             refreshBooks()
             resetForm()
             setValues({
@@ -103,20 +104,10 @@ const AddBook = () => {
                 startDate: null,
                 endDate: null
             });
-            // window.scrollTo({ top: 0, behaviour: 'smooth' });
-            el.scrollIntoView({ behavior: 'smooth' })
         } catch (error) {
-            // window.scrollTo({ top: 0, behaviour: 'smooth' });
-            console.log(error)
-            el.scrollIntoView({ behavior: 'smooth' })
             setStatus({ response: error.response.data.message })
         }
-    }, [refreshBooks])
-
-    const handleScroll = () => {
-        const el = document.querySelector('#top')
-        el.scrollIntoView({ behavior: 'smooth' })
-    }
+    }, [refreshBooks, location.pathname, navigate])
 
     return (
         <div className="p-4">
@@ -132,7 +123,6 @@ const AddBook = () => {
                         <Form>
                             <TextField name="title" label="Title" />
                             <TextField name="author" label="Author" />
-                            {/* <ScrollToError /> */}
                             <div id="start" className="text-md font-semibold text-red-500">{status?.response}</div>
 
                             <p id="status-group" className="font-semibold">Status</p>
@@ -144,7 +134,6 @@ const AddBook = () => {
                                 <RadioButton name="status" value="to read"
                                     label="To read" />
                             </div>
-
 
                             {values.status === "read" &&
                                 <>
@@ -197,8 +186,9 @@ const AddBook = () => {
 
                             <button type="submit"
                                 className="px-4 py-2 mt-2 text-center bg-lighter-accent hover:bg-main-accent-hover text-light-bg font-semibold rounded-md"
-                                // onClick={ handleScroll}
                             >Add book</button>
+
+                            <ScrollToTop />
                         </Form>
                     )
                 }}
