+++ src/pages/user/AddBookPage/components/ScrollToTop.jsx
@@ -1,0 +1,17 @@
+import { useFormikContext } from "formik"
+import { useEffect } from "react"
+
+const ScrollToTop = () => {
+    const formik = useFormikContext()
+    const submitting = formik?.isSubmitting;
+
+    useEffect(() => {
+        const el = document.querySelector('h1')
+        setTimeout(() => {
+            el.scrollIntoView({ behavior: 'smooth' })
+        }, 1000);
+    }, [submitting])
+
+    return null
+}
+export default ScrollToTop
